+++ src/shared/src/types.rs
@@ -1,5 +1,4 @@
 use candid::{CandidType, Deserialize, Principal};
-use ic_canister_sig_creation::{extract_raw_root_pk_from_der, IC_ROOT_PK_DER};
 use std::fmt::Debug;
 
 pub type Timestamp = u64;
@@ -41,33 +40,6 @@ pub struct Config {
     pub supported_credentials: Option<Vec<SupportedCredential>>,
     /// Root of trust for checking canister signatures.
     pub ic_root_key_raw: Option<Vec<u8>>,
-}
-
-impl From<InitArg> for Config {
-    /// Creates a new `Config` from the provided `InitArg`.
-    ///
-    /// # Panics
-    /// - If the root key cannot be parsed.
-    fn from(arg: InitArg) -> Self {
-        let InitArg {
-            ecdsa_key_name,
-            allowed_callers,
-            supported_credentials,
-            ic_root_key_der,
-        } = arg;
-        let ic_root_key_raw = match extract_raw_root_pk_from_der(
-            &ic_root_key_der.unwrap_or_else(|| IC_ROOT_PK_DER.to_vec()),
-        ) {
-            Ok(root_key) => root_key,
-            Err(msg) => panic!("{}", format!("Error parsing root key: {msg}")),
-        };
-        Config {
-            ecdsa_key_name,
-            allowed_callers,
-            supported_credentials,
-            ic_root_key_raw: Some(ic_root_key_raw),
-        }
-    }
 }
 
 pub mod transaction {
